LayoutTests/fast/dom/HTMLTemplateElement/cycles-expected.txt

Test that cycles are not allowed in template content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS template.content.appendChild(template) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS template.content.appendChild(outerDiv) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS innerDiv.appendChild(template) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS innerDiv.appendChild(outerDiv) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS innerTemplate.appendChild(outerDiv) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS successfullyParsed is true

TEST COMPLETE
